+++ R/initialize.R
@@ -789,6 +789,7 @@ augment_iea_df <- function(.iea_df,
 #' Default argument values assume that `rename_iea_df_cols()` has been called on `.iea_df`.
 #'
 #' @param .iea_df a IEA data frame whose columns have been renamed by `rename_iea_df_cols()`
+#' @param col_names a list of column names in IEA data frames. Default is `IEATools::iea_cols`.
 #' @param year the name of the year column created in `.iea_df` by this function. (Default is "Year".)
 #' @param method the name of the method column created in `.iea_df` by this function. (Default is "Method".)
 #' @param last_stage the name of the last stage column created in `.iea_df` by this function. (Default is "Last.stage".)